+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import CategoryModel, { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -60,6 +61,9 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Failed find one category");
+      }
       const result = await CategoryModel.findById(id);
       if (!result) {
         return response.notFound(res, "Category not found");
@@ -72,6 +76,9 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+       if (!isValidObjectId(id)) {
+         return response.notFound(res, "Failed update category");
+       }
       const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -83,6 +90,9 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+       if (!isValidObjectId(id)) {
+         return response.notFound(res, "Failed remove category");
+       }
       const result = await CategoryModel.findByIdAndDelete(id, {
         new: true,
       });
